backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.11
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mysql-connector-python==9.2.0
numpy==2.2.4
ollama==0.4.7
pgvector==0.4.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
